currency: $500 wallet -- widen Item.amount to ushort; throwables stack to 5
strawberry: "wire up the $500 cap. int, not float. no pennies here. ... make
frag grenades, smokes, flares stack to 5."

`Item.amount` byte -> ushort. It had to move because for money `amount` IS
the dollar figure -- that identity is the whole carrier design -- so $500
simply does not fit in a byte. Integer, never float: money is whole dollars
and nothing about a stack count wants a fraction.

USHORT rather than int, which is a judgement call worth stating since the ask
said "int": 65535 is headroom for any stack we will ever have, and it costs
the wire 2 bytes per item instead of 4. Still an integer type, no floats
anywhere near it. Say the word if you want the full int.

Protocol: amount now crosses as UInt16 (inventory + world-item replication,
both directions) and mixes into the desync hash as UInt32. Saves are JSON, so
a widened number is backward compatible -- old saves with values <= 255 load
unchanged and no migration is needed.

ReloadSwapCommand.SpentAmount widened too. A magazine never fills a byte, so
this one is not strictly required, but a quantity that silently truncates on
the wire is not worth the byte it saves.

Items.AddCore: a declared stackSize now always wins, even under
StackingEnabled. That option means "stack more freely", so it must not be the
thing that clips money back to the old 255 ceiling.

Throwables stack to 5, driven off Throwables.All rather than a hand-written
id list -- the table already knows which ids are throwables (both grenades,
flares 255-260, smoke 261-268), so one added later stacks without anyone
remembering this line, and no id can be in one list but not the other.

Tests: the spill test is now DERIVED from the ceiling instead of written
against it. It was "$300 overflows 255"; at 500 that literal still passes
while testing no overflow at all, which is exactly the way a test goes quietly
vacuous. It now spends MaxPerStack/100+1 hundreds, so it keeps meaning what
it says wherever the ceiling lands. The widest-fan test is a deliberate
tripwire at 8 notes ($485): moving the ceiling SHOULD break it, because that
is the prompt to re-check the icon layout, not a number to quietly bump.

+++ core/UnturnedNet/NetWorldHost.cs
@@ -1242,7 +1242,7 @@ namespace UnturnedGodot.Net
         public bool SendConsume(byte page, byte x, byte y)
             => SendCommand(ReplicationIds.CommandConsume, new ConsumeCommand { Page = page, X = x, Y = y }.Write);
 
-        public bool SendReloadSwap(byte page, byte x, byte y, ushort spentId, byte spentAmount)
+        public bool SendReloadSwap(byte page, byte x, byte y, ushort spentId, ushort spentAmount)
             => SendCommand(ReplicationIds.CommandReloadSwap, new ReloadSwapCommand { Page = page, X = x, Y = y, SpentId = spentId, SpentAmount = spentAmount }.Write);
 
         public bool SendGunUnload(byte page, byte x, byte y, ushort roundId, byte count)
